Second Commit
Uploading updated files as well as new work done. This also includes the updated version of Report1 Deep Learning v8. It should actually be called v9 but the whole point of github is that we don't have to do that. Let's see how this goes. This would also be my first file for which I "push" or commit a new version.